% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs.data.r
\docType{data}
\name{hs.data}
\alias{hs.data}
\title{Heshun human population}
\description{
The \code{hs.data} dataset consists of three data frames:
\itemize{
\item\code{hs.data[[1]]}: Human population of 167 sampling villages within the Heshun region, China, which serves as the sampling layer. The sampling attribute is \code{Population}.
\item\code{hs.data[[2]]}: A zonation map with 5 strata that serves as a candidate SSH layer.
\item\code{hs.data[[3]]}: The administrative division of 10 townships in Heshun county that served as the reporting layer.}
}
\examples{
data(hs.data)
}
\references{
Li, L., Wang, J., & Wu, J. (2012). A spatial model to predict the incidence of neural tube defects. \emph{BMC Public Health}, 12(1), 951.
}
