% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.plot.ci.r
\name{sandwich.plot.ci}
\alias{sandwich.plot.ci}
\title{Plot confidence interval for the mean}
\usage{
sandwich.plot.ci(object)
}
\arguments{
\item{object}{An \code{sf} object generated from the \code{sandwich.ci} function.}
}
\description{
\code{sandwich.plot.ci} plots the lower and upper bounds of the confidence interval for the mean value of each reporting unit.
}
