% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.data.r
\docType{data}
\name{sim.data}
\alias{sim.data}
\title{Simulated data}
\description{
The \code{sim.data} dataset consists of three \code{sf} objects:
\itemize{
\item\code{sim.data[[1]]}: 41 sampling units. The sampling attribute is \code{Value}.
\item\code{sim.data[[2]]}: A zonation map with 4 strata that serves as a candidate SSH layer.
\item\code{sim.data[[3]]}: 7 reporting units.}
}
\examples{
data(sim.data)
}
