% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.data.txt.r
\name{ssh.data.txt}
\alias{ssh.data.txt}
\title{Prepare text data for Geodetector}
\usage{
ssh.data.txt(object)
}
\arguments{
\item{object}{A list of 2 data frames generated by \code{load.data.txt}, including a file linking sampling and SSH layers linking reporting and SSH layers and a file.}
}
\description{
\code{ssh.data.txt} prepare the input from text files for the geographical detector model.
}
\seealso{
\link[https://cran.r-project.org/package=geodetector]{geodetector R package}
}
