% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sandwich.r
\name{summary}
\alias{summary}
\alias{summary.sandwich.model}
\title{Summarize interpolation results}
\usage{
\method{summary}{sandwich.model}(object, ...)
}
\arguments{
\item{object}{A \code{sandwich.model} object generated by \code{\link{sandwich.model}}.}

\item{...}{Ignored.}
}
\value{
Summarized statistics for the estimated mean and standard error.
}
\description{
Summarize the estimated mean and standard error.
}
\examples{
data(sim.data)
sim.sw <- sandwich.model(object=sim.data, sampling.attr="Value", type="shp")
summary(object=sim.sw)

}
\seealso{
\code{\link{sandwich.model}}
}
