% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.ci.r
\name{sandwich.ci}
\alias{sandwich.ci}
\title{Calculate confidence interval for the mean}
\usage{
sandwich.ci(object, level=.95)
}
\arguments{
\item{object}{An \code{sf} object or a data frame generated from the \code{sandwich.model} function.}

\item{level}{The confidence level required. By default, \code{level} = .95.}
}
\description{
\code{sandwich.ci} calculates the confidence interval for the mean value of each reporting unit.
}
\examples{
data(hs.data)
hs.sw <- sandwich.model(object=hs.data, sampling.attr="Population", type="shp")
hs.sw.ci <- sandwich.ci(object=hs.sw, level=.95)

}
