% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.model.r
\name{sandwich.model}
\alias{sandwich.model}
\title{Perform Sandwich model-based mapping}
\usage{
sandwich.model(object, sampling.attr, type="shp", ssh.id.col=NULL, ssh.weights=NULL)
}
\arguments{
\item{object}{When \code{type="shp"}, \code{object} is a list of 3 \code{sf} objects generated by \code{load.data.shp}, including a POINT \code{sf} object used as the sampling layer, a POLYGON \code{sf} object used as the SSH layer, and a POLYGON \code{sf} object used as the SSH layer. When \code{type="txt"}, \code{object} is a list of 2 data frames generated by \code{load.data.txt}, including a file linking sampling and SSH layers linking reporting and SSH layers and a file.}

\item{sampling.attr}{A \code{string} denoting the name of the attribute in the sampling layer to be interpolated.}

\item{type}{A \code{string} denoting the type of input data. \code{type="shp"} denotes shapefiles, and \code{type="txt"} denotes text files. By default, \code{type="shp"}.}

\item{ssh.id.col}{A \code{string} denoting the column that specifies which stratum each sampling unit falls into in the file linking sampling and SSH layers.}

\item{ssh.weights}{A \code{list} that specifies the strata in the SSH layer and their corresponding columns of weights in the file linking reporting and SSH layers.}
}
\description{
\code{sandwich.model} estimate the mean value of an attribute and its standard error for each reporting unit using the sandwich estimation framework by Wang et al. (2013).
}
\examples{
data(hs.data)
hs.sw <- sandwich.model(object=hs.data, sampling.attr="Population", type="shp")

}
\references{
Wang, J. F., Haining, R., Liu, T. J., Li, L. F., & Jiang, C. S. (2013). Sandwich estimation for multi-unit reporting on a stratified heterogeneous surface. \emph{Environment and Planning A}, 45(10), 2515-2534. doi:\link[https://doi.org/10.1068/a44710]{10.1068/a44710}
}
