% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.plot.mean.r
\name{sandwich.plot.mean}
\alias{sandwich.plot.mean}
\title{Plot Sandwich estimates}
\usage{
sandwich.plot.mean(object)
}
\arguments{
\item{object}{An \code{sf} object generated from the \code{sandwich.model} function.}
}
\description{
\code{sandwich.plot.mean} plots the mean value of an attribute for each reporting unit estimated by the Sandwich model.
}
\examples{
data(hs.data)
hs.sw <- sandwich.model(object=hs.data, sampling.attr="Population", type="shp")
sandwich.plot.mean(object=hs.sw)

}
\references{
Wang, J. F., Haining, R., Liu, T. J., Li, L. F., & Jiang, C. S. (2013). Sandwich estimation for multi-unit reporting on a stratified heterogeneous surface. \emph{Environment and Planning A}, 45(10), 2515-2534. doi:\link[https://doi.org/10.1068/a44710]{10.1068/a44710}
}
