% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.data.txt.r
\name{ssh.data.txt}
\alias{ssh.data.txt}
\title{Prepare text data for Geodetector}
\usage{
ssh.data.txt(object)
}
\arguments{
\item{object}{A list of two data frames generated by \code{\link{load.data.txt}}, including a file linking sampling and SSH layers linking reporting and SSH layers and a file.}
}
\value{
A list of two data frames.
}
\description{
Prepare the input from text files for the geographical detector model.
}
\examples{
data(bc.data)
bc.join <- ssh.data.txt(object=bc.data)
}
